added controller + caching for secrets

+++ src/Api/Controllers/UserController.cs
@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MyWebApi.App.DTO;
 using MyWebApi.App.Interfaces;
@@ -7,15 +8,26 @@ namespace MyWebApi.Api.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class UsersController : ControllerBase
+[Authorize]
+public class UsersController(ICrudService<CreateUserDto, UserDto> service, UserContext userContext) : ControllerBase
 {
-    private readonly ICrudService<CreateUserDto, UserDto> _service;
+    private readonly ICrudService<CreateUserDto, UserDto> _service = service;
+    private readonly UserContext _userContext = userContext;
 
-    public UsersController(ICrudService<CreateUserDto, UserDto> service)
+
+
+    // later I would add roles here, to make sure only admin can check etc.
+    // left it unsafe for demo atm
+    [HttpGet("{id:guid}")]
+    public ActionResult<UserDto> GetById(Guid id)
     {
-        _service = service;
+        var userId = _userContext.UserId;
+        if (userId is null) return Unauthorized();        
+        UserDto user = _service.GetById(id);
+        return Ok(user);
     }
 
+    
     [HttpGet]
     public ActionResult<PagedResult<UserDto>> GetUsers([FromQuery] int page = 1, [FromQuery] int pageSize = 10)
     {
@@ -23,14 +35,9 @@ public class UsersController : ControllerBase
         return Ok(result);
     }
 
-    [HttpGet("{id:guid}")]
-    public ActionResult<UserDto> GetById(Guid id)
-    {
-        UserDto user = _service.GetById(id);
-        return Ok(user);
-    }
 
     [HttpPost]
+    [AllowAnonymous]
     public IActionResult Create([FromBody] CreateUserDto dto)
     {
         _service.Add(dto);
@@ -40,6 +47,9 @@ public class UsersController : ControllerBase
     [HttpPut("{id:guid}")]
     public ActionResult<UserDto> Update(Guid id, [FromBody] CreateUserDto dto)
     {
+        var userId = _userContext.UserId;
+        if(userId is null) return Unauthorized();
+        if(userId != id) return Forbid();
         var updated = _service.Update(id, dto);
         return Ok(updated);
     }
@@ -47,6 +57,9 @@ public class UsersController : ControllerBase
     [HttpDelete("{id:guid}")]
     public IActionResult Delete(Guid id)
     {
+        var userId = _userContext.UserId;
+        if(userId is null) return Unauthorized();
+        if(userId != id) return Forbid();
         _service.Delete(id);
         return NoContent();
     }
